+++ src/Awesomite/ErrorDumper/StandardExceptions/FatalErrorException.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Awesomite\ErrorDumper\StandardExceptions;
-
-/**
- * @internal
- */
-class FatalErrorException extends ErrorException {}